requirements.txt
# MCP (Model Context Protocol)
mcp>=1.0.0

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & Vector Database
openai>=1.3.0
chromadb>=0.4.18

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Caching & Performance
redis>=5.0.0
hiredis>=2.2.3

# Payment Processing
stripe>=7.4.0

# Voice Processing
twilio>=8.10.0
websockets>=12.0
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0
webrtcvad>=2.0.10
aiortc>=1.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7
prometheus-client>=0.19.0

# Security & Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
tenacity>=8.2.3
python-dotenv>=1.0.0
